Support valid DOIs ending in punctuation
During some internal data clean-up work at @altmetric, we discovered
DOIs that end with closing parentheses. These were being dropped by our
extraction regular expression as we enforce a word boundary (which
splits when a word character is followed by a non-word character such as
a closing parenthesis).

By removing the strict word boundary, we can now pick up these closing
parentheses and filter out any unwanted trailing punctuation in a second
pass over the matches.

We do this by explicitly dropping the last character if it is
punctuation but _not_ if the DOI ends in a single, balanced
parenthetical group of characters or if it is an early Wiley DOI (which
end in a certain sequence of punctuation).

This means that we can successfully extract valid DOIs such as
10.1002/0471221929.ch26(v) but not invalid ones such as
10.1007/978-3-319-12580-0).