Don't extract DOIs with pure punctuation suffixes
As we now mutate DOIs after extracting them, it is possible for us to
invalidate the DOI during extraction. The simplest example of this is a
DOI with a suffix made entirely of trailing punctuation.

To fix this, ensure that we always revalidate a DOI when trying to strip
off any trailing punctuation.