Refactor ISBN class to a more imperative style
The use of array_filter, array_map and various callables in the ISBN
class was starting to obstruct readability so let's abandon the mix of
programming styles in favour of plain, old imperative programming.

This means dropping map and filter for explicit for loops which moves us
away from describing _what_ to do and more toward _how_ to do things but
I suspect being more consistent and predictable is better than
shoehorning different styles in one class.